+++ component/worldMenu/subComponent/getRowContainerElement.js
@@ -1,0 +1,7 @@
+const getRowContainerElement = () => {
+	const rowContainer1 = document.getElementById("rowId1");
+	const rowContainer2 = document.getElementById("rowId2");
+	return { rowContainer1, rowContainer2 };
+};
+
+export default getRowContainerElement;
